% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferentialSampling.R
\name{burnin_after}
\alias{burnin_after}
\title{Apply Burnin to MCMC Samples}
\usage{
burnin_after(output, n.burn)
}
\arguments{
\item{output}{List. Output of the preferentialSampling.}

\item{n.burn}{Numeric. Number of samples to discard in burnin.}
}
\value{
List. Identical to the output of preferentialSampling but with truncated posterior samples.
}
\description{
This function truncates the first n.burn posterior samples of the
output of the preferentialSampling function.
}
