% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{calc_log_odds_pr}
\alias{calc_log_odds_pr}
\title{calc_log_odds_pr}
\usage{
calc_log_odds_pr(sampling, prevalence)
}
\arguments{
\item{sampling}{}

\item{prevalence}{}
}
\value{

}
\description{
calculates the log odds from the poisson regression models
}
