% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{calc_log_odds_sp}
\alias{calc_log_odds_sp}
\title{calc_log_odds_sp}
\usage{
calc_log_odds_sp(outputs, sampling, prevalence)
}
\arguments{
\item{outputs}{}

\item{sampling}{}

\item{prevalence}{}
}
\value{

}
\description{
calculates the log odds from the spatial poisson models
}
