% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_w}
\alias{combine_w}
\title{combine_w}
\usage{
combine_w(w_est, w_krige, status)
}
\arguments{
\item{w_est}{estimated random effects of observed sites}

\item{w_krige}{kriged random effects}

\item{status}{vector of TRUE/FALSE values indicating observation}
}
\value{
vector of combined estimated and kriged random effects
}
\description{
Combines estimated and kriged random effects 
according to cell observation status.
}
