% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferentialSampling.R
\name{continue_mcmc}
\alias{continue_mcmc}
\title{Continue running MCMC}
\usage{
continue_mcmc(data, D, output, n.sample)
}
\arguments{
\item{data}{List. Input data for the model.}

\item{D}{Matrix. distance matrix describing grid cells in study region.}

\item{output}{List. Output of the preferentialSampling function.}

\item{n.sample}{Numeric. Number of additional MCMC samples to generate.}
}
\value{
a list of posterior samples and associated quantities
}
\description{
This function continues to run the MCMC sampling routine for the 
preferentialSampling model.
}
