% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.R
\name{K}
\alias{K}
\title{K}
\usage{
K(p)
}
\arguments{
\item{p}{numeric. vector of momenta.}
}
\value{
numeric. kinetic energy.
}
\description{
Returns the kinetic energy of a momentum vector. Used as an auxiliary
function in Hamiltonian Monte Carlo samplers.
}
\examples{

momentum <- rnorm(20)
K(momentum)

}
