% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{load_output}
\alias{load_output}
\title{load_output}
\usage{
load_output(
  fname,
  src = "/Users/brianconroy/Documents/research/dataInt/output/"
)
}
\arguments{
\item{file}{(character) name of output JSON file to be loaded}
}
\value{
(list) loaded output
}
\description{
load_output
}
