% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overlay}
\alias{overlay}
\title{overlay}
\usage{
overlay(vals, r)
}
\arguments{
\item{vals}{vector of ordered cell values}

\item{r}{raster}
}
\value{
new raster with replaced values
}
\description{
overlay
}
