% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{plot_traces}
\alias{plot_traces}
\title{plot_traces}
\usage{
plot_traces(outputs, sampling, prevalence)
}
\arguments{
\item{outputs}{(list) mcmc preferential sampling outputs}

\item{sampling}{(character) strength of preferential sampling}

\item{prevalence}{(character) disease prevalence}
}
\value{

}
\description{
plots traceplots for the preferential sampling model
}
