% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{save_output}
\alias{save_output}
\title{save_output}
\usage{
save_output(
  output,
  fname,
  dst = "/Users/brianconroy/Documents/research/dataInt/output/"
)
}
\arguments{
\item{output}{(list) output of an MCMC function}

\item{fname}{(character) file name}
}
\value{
writes output as json file
}
\description{
save_output
}
