% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_param}
\alias{summarize_param}
\title{summarize_param}
\usage{
summarize_param(name, true, estimated)
}
\arguments{
\item{name}{parameter name}

\item{true}{true parameter value}

\item{estimated}{estimated parameter value}
}
\value{
list summarizing estimate and bias
}
\description{
summarize_param
}
