% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simUtils.R
\name{table_params}
\alias{table_params}
\title{table_params}
\usage{
table_params(outputs, sampling, prevalence)
}
\arguments{
\item{outputs}{(list) mcmc outputs}

\item{sampling}{(character) strength of sampling}

\item{prevalence}{(character) disease prevalence ("low", "medium", "high")}
}
\value{
(list) list of lists containing parameter
estimates, true values and biases of different models
}
\description{
table_params
}
