% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_tr}
\alias{view_tr}
\title{view_tr}
\usage{
view_tr(samples, trueval = NULL, title = "", ylim = NULL)
}
\arguments{
\item{samples}{vector of mcmc samples}

\item{trueval}{optional true parameter value displayed as horizontal line}

\item{title}{optional plot title}

\item{ylim}{optional y axis limits}
}
\value{

}
\description{
view_tr
}
