% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_tr_w}
\alias{view_tr_w}
\title{view_tr_w}
\usage{
view_tr_w(samples, w_true = NULL, page = 1)
}
\arguments{
\item{samples}{matrix of mcmc samples}

\item{w_true}{optional vector of true random effect values}

\item{page}{view offset}
}
\value{

}
\description{
multipanel (4x4) view of random effect traceplots
}
